Removed the option to write to a separate output file.
This idea was at odds with the rest of the structure of the program. To facilitate subcommands like renaming or converting file formats, we had to make exceptions if the file had a separate output. The same goes for the backup system. This extra complexity was not worth it; output files were only even valid in certain situations.

As part of this restructuring, the backup system has been improved. It now supports backing up of file renames/moves.