Big refactor to make subcommands process an array of files which are now only loading from file when they need to be.
This removes the need for the complicated workarounds where something like the Normaliser need to read all the files first before processing it. And now as audio data is only loaded from file when GetWritableAudio is called on the InputAudioFile, calls to signet that only rename or move files do not need to read and write which makes thing much quicker.