Improve pitch detection even further
To achieve this, the suitability of a detected pitch is now based also on the RMS of the chunk, where louder chunks have a slight preference. Also the chunk comparison is done with a curve rather than a cutoff boolean.

As well as this, the pitch detector now checks the pitch at multiple octaves and works out the pitch based off of all of the results. This reduces the error caused by the pitch detection algorithm failing for very high or low pitched audio.