Read/save all metadata into flac files using a custom id "SGNT"
The library cereal was added for easy serialisation of an audio file's metadata into JSON. We then save this JSON into a custom chunk in the FLAC file, using the id "SGNT". This means data will not be lost when using signet multiple times on an audio file or from converting back and forth between wav and flac. It also means other software can also look for Signet's added data and use it themselves.